requirements.txt
accelerate>=1.2.0
matplotlib>=3.8.0
numpy>=1.26.0
pillow>=10.4.0
requests>=2.32.0
safetensors>=0.4.5
sentencepiece>=0.2.0
torch>=2.4.0
torchvision>=0.19.0
transformers>=4.51.0
